+++ xalpha/__init__.py
@@ -11,6 +11,6 @@ from xalpha.multiple import mul, mulfix, imul
 from xalpha.realtime import rfundinfo, review
 from xalpha.record import record, irecord
 from xalpha.trade import trade, itrade
-from xalpha.universal import get_daily, get_rt, get_bar, set_backend, set_handler
+from xalpha.universal import get_daily, get_rt, get_bar, set_backend, set_handler, vinfo
 from xalpha.provider import show_providers, set_proxy
 from xalpha.toolbox import PEBHistory, Compare, SWPEBHistory, QDIIPredict, set_holdings